python-llm/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.14
langchain-openai>=0.2,<0.3
langchain-community>=0.3,<0.4
python-dotenv